Implement `humboldt configure`
The structure of Thor did not allow the configuration file values to be
injected into options with higher priority than the default, but lower
priority than the command line arguments. Therefor, the default
functionality of Thor is no longer used.

`humboldt configure` will create a .humboldt.yml file in the current
directory, and provides a simple interface to set some defaults.