Add Java-versions of the basic type converters
The Json/Encoded subclasses remain in Ruby, but the Java-library still
mentions these classes as they need to have their Java-class redefined,
as it otherwise cannot be converted to its superclass. This would not
be necessary if the Java-classes were defined prior to their Ruby
counterparts. However, that would not be desirable, as the last method
definition is kept.