Use a Library instead of a Service for JRuby extension
I'm not sure why, but I think it might have something to do with the
humboldt jar being loaded from outside the JRuby classloader, meaning
that either the extension isn't magically started at all (or it is
started before the humboldt Ruby code has executed).

Instead of trying to figure this out, I'm going with the more direct
approach and explicitly loading the Java-extensions from the Ruby file.