Add type conversion for user-defined packable values
This adds type conversion for any class including the Packable module,
which in essence just forces the implementation of two methods
`#to_packable` and `.from_packable`. These methods are used when
de-/serializing the packable object to and from MessagePack.

Two handy helper classes `Key` and `Value` allows for creating
Struct-like classes which are packable.