Add humbold.BinaryComparator for using with Hadoop's BinaryPartitioner
This replaces Humboldt::BinaryPrefixPartitioner and friends, if you can live with having to write a more cryptic job configuration. The idea is that Hadoop's BinaryPartitioner is both more efficient and more versatile than what's currently in Humboldt, but Hadoop is missing a comparator that works with the same offset configuration.

For secondary sort instead of

    partitioner(MyPartitionerBasedOnBinaryPrefixPartitioner)
    grouping_comparator(MyComparatorBasedOnBinaryPrefixComparator)

you'll have to do (assuming the classes in the example above are configured to look at the first 12 bytes)

    raw do |job|
      job.set_partitioner_class(Hadoop::Mapreduce::Lib::Partition::BinaryPartitioner)
      Hadoop::Mapreduce::Lib::Partition::BinaryPartitioner.set_offsets(job.configuration, 0, 11)
      job.set_grouping_comparator_class(Humboldt::JavaLib::BinaryComparator)
      Humboldt::JavaLib::BinaryComparator.set_offsets(job.configuration, 0, 11)
    end

if you want to do secondary sort on a 12 byte suffix you can use -12 and -1 as the left and right offset for both the partitioner and comparator. The offsets are somehow based on Python string slicing, which I guess is more or less how Ruby's range slicing works too (it's just that you usually don't use ranges for slicing strings in Ruby). This means that if you want the first 12 bytes that's 0, 11 and not 0, 12 as you might expect if you're like me and assume the second argument is the length.